Delete Managers/LevelSelection.cs
Transfer LevelSelection.cs content to GameManager.cs in order to have one canvas in the entire game. The reason being that we would prefer that the canvas’ render mode be set to ‘Screen Space - Camera’ in order to set its sorting layer behind certain game objects that have a sprite renderer attached.